// src/types/generated/v600.ts
export interface Token {
    supply: bigint
    cap?: (TokenCap | undefined)
    freezeState?: (FreezeState | undefined)
    minimumBalance: bigint
    sufficiency: Sufficiency
    mintDeposit: bigint
    attributeCount: number
    marketBehavior?: (TokenMarketBehavior | undefined)
    listingForbidden: boolean
    metadata: DefaultTokenMetadata
}

export type DefaultTokenMetadata = DefaultTokenMetadata_Foreign | DefaultTokenMetadata_Native

export interface DefaultTokenMetadata_Foreign {
    __kind: 'Foreign'
    value: DefaultForeignTokenMetadata
}

export interface DefaultTokenMetadata_Native {
    __kind: 'Native'
}

export interface DefaultForeignTokenMetadata {
    decimalCount: number
    name: BoundedString
    symbol: Bytes
    location?: (V3MultiLocation | undefined)
    unitsPerSecond?: (bigint | undefined)
    premintedSupply: bigint
}
// ...
export type TokenMarketBehavior = TokenMarketBehavior_HasRoyalty | TokenMarketBehavior_IsCurrency

export interface TokenMarketBehavior_HasRoyalty {
    __kind: 'HasRoyalty'
    value: DefaultRoyalty
}

export interface TokenMarketBehavior_IsCurrency {
    __kind: 'IsCurrency'
}

export interface DefaultRoyalty {
    beneficiary: AccountId32
    percentage: number
}

export type AccountId32 = Bytes

export type Sufficiency = Sufficiency_Insufficient | Sufficiency_Sufficient

export interface Sufficiency_Insufficient {
    __kind: 'Insufficient'
    unitPrice: bigint
}

export interface Sufficiency_Sufficient {
    __kind: 'Sufficient'
}

export type FreezeState = FreezeState_Never | FreezeState_Permanent | FreezeState_Temporary

export interface FreezeState_Never {
    __kind: 'Never'
}

export interface FreezeState_Permanent {
    __kind: 'Permanent'
}

export interface FreezeState_Temporary {
    __kind: 'Temporary'
}

export type TokenCap = TokenCap_CollapsingSupply | TokenCap_SingleMint | TokenCap_Supply

export interface TokenCap_CollapsingSupply {
    __kind: 'CollapsingSupply'
    value: bigint
}

export interface TokenCap_SingleMint {
    __kind: 'SingleMint'
}

export interface TokenCap_Supply {
    __kind: 'Supply'
    value: bigint
}
// ...
export const DefaultTokenMetadata: sts.Type<DefaultTokenMetadata> = sts.closedEnum(() => {
    return  {
        Foreign: DefaultForeignTokenMetadata,
        Native: sts.unit(),
    }
})

export const DefaultForeignTokenMetadata: sts.Type<DefaultForeignTokenMetadata> = sts.struct(() => {
    return  {
        decimalCount: sts.number(),
        name: BoundedString,
        symbol: sts.bytes(),
        location: sts.option(() => V3MultiLocation),
        unitsPerSecond: sts.option(() => sts.bigint()),
        premintedSupply: sts.bigint(),
    }
})

export const BoundedString = sts.bytes()

export const Type_145: sts.Type<Type_145> = sts.closedEnum(() => {
    return  {
        NoMutation: sts.unit(),
        SomeMutation: sts.boolean(),
    }
})

export const Type_142: sts.Type<Type_142> = sts.closedEnum(() => {
    return  {
        NoMutation: sts.unit(),
        SomeMutation: sts.option(() => TokenMarketBehavior),
    }
})

export const TokenMarketBehavior: sts.Type<TokenMarketBehavior> = sts.closedEnum(() => {
    return  {
        HasRoyalty: DefaultRoyalty,
        IsCurrency: sts.unit(),
    }
})

export const DefaultRoyalty: sts.Type<DefaultRoyalty> = sts.struct(() => {
    return  {
        beneficiary: AccountId32,
        percentage: sts.number(),
    }
})
// ...
export const FreezeState: sts.Type<FreezeState> = sts.closedEnum(() => {
    return  {
        Never: sts.unit(),
        Permanent: sts.unit(),
        Temporary: sts.unit(),
    }
})
// ...
export const TokenCap: sts.Type<TokenCap> = sts.closedEnum(() => {
    return  {
        CollapsingSupply: sts.bigint(),
        SingleMint: sts.unit(),
        Supply: sts.bigint(),
    }
})
// ...
export const Token: sts.Type<Token> = sts.struct(() => {
    return  {
        supply: sts.bigint(),
        cap: sts.option(() => TokenCap),
        freezeState: sts.option(() => FreezeState),
        minimumBalance: sts.bigint(),
        sufficiency: Sufficiency,
        mintDeposit: sts.bigint(),
        attributeCount: sts.number(),
        marketBehavior: sts.option(() => TokenMarketBehavior),
        listingForbidden: sts.boolean(),
        metadata: DefaultTokenMetadata,
    }
})

export const Sufficiency: sts.Type<Sufficiency> = sts.closedEnum(() => {
    return  {
        Insufficient: sts.enumStruct({
            unitPrice: sts.bigint(),
        }),
        Sufficient: sts.unit(),
    }
})